+++ include/fiwix/statfs.h
@@ -9,20 +9,20 @@
 #define _FIWIX_STATFS_H
 
 typedef struct {
-	long int val[2];
+	int val[2];
 } fsid_t;
 
 struct statfs {
-	long int f_type;
-	long int f_bsize;
-	long int f_blocks;
-	long int f_bfree;
-	long int f_bavail;
-	long int f_files;
-	long int f_ffree;
+	int f_type;
+	int f_bsize;
+	int f_blocks;
+	int f_bfree;
+	int f_bavail;
+	int f_files;
+	int f_ffree;
 	fsid_t f_fsid;
-	long int f_namelen;
-	long int f_spare[6];
+	int f_namelen;
+	int f_spare[6];
 };
 
 #endif /* _FIWIX_STATFS_H */
